# docs/modbus_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API文档",
        "title": "物联大师网关接口文档",
        "contact": {},
        "version": "1.0 版本"
    },
    "basePath": "/api/gateway/api/",
    "paths": {
...
        "/serial/count": {
            "post": {
                "description": "查询串口数量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "serial"
                ],
                "summary": "查询串口数量",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ParamSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-int64"
                        }
                    }
                }
            }
        },
        "/serial/create": {
            "post": {
                "description": "创建串口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "serial"
                ],
                "summary": "创建串口",
                "parameters": [
                    {
                        "description": "串口信息",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Serial"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Serial"
                        }
                    }
                }
            }
        },
...
        "/serial/import": {
            "post": {
                "description": "导入串口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "导入串口",
                "parameters": [
                    {
                        "type": "file",
                        "description": "压缩包",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-int64"
                        }
                    }
                }
            }
        },
        "/serial/list": {
            "get": {
                "description": "查询串口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "serial"
                ],
                "summary": "查询串口",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "skip",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyList-types_Serial"
                        }
                    }
                }
            }
        },
...
        "/serial/search": {
            "post": {
                "description": "查询串口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "serial"
                ],
                "summary": "查询串口",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ParamSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyList-types_Serial"
                        }
                    }
                }
            }
        },
        "/serial/{id}": {
            "get": {
                "description": "获取串口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "serial"
                ],
                "summary": "获取串口",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "串口ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Serial"
                        }
                    }
                }
            },
            "post": {
                "description": "修改串口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "serial"
                ],
                "summary": "修改串口",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "串口ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "串口信息",
                        "name": "serial",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Serial"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Serial"
                        }
                    }
                }
            }
        },
        "/serial/{id}/delete": {
            "get": {
                "description": "删除串口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "serial"
                ],
                "summary": "删除串口",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "串口ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Serial"
                        }
                    }
                }
            }
        },
        "/serial/{id}/disable": {
            "get": {
                "description": "禁用串口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "serial"
                ],
                "summary": "禁用串口",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "串口ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Serial"
                        }
                    }
                }
            }
        },
        "/serial/{id}/enable": {
            "get": {
                "description": "启用串口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "serial"
                ],
                "summary": "启用串口",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "串口ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Serial"
                        }
                    }
                }
            }
        },
        "/serial/{id}/start": {
            "get": {
                "description": "启动串口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "serial"
                ],
                "summary": "启动串口",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "串口ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Serial"
                        }
                    }
                }
            }
        },
        "/serial/{id}/stop": {
            "get": {
                "description": "停止串口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "serial"
                ],
                "summary": "停止串口",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "串口ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-types_Serial"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ParamSearch": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "object",
                    "additionalProperties": true
                },
                "keyword": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "skip": {
                    "type": "integer"
                },
                "sort": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        },
        "api.ReplyData-int64": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "api.ReplyData-types_Serial": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.Serial"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "api.ReplyList-types_Serial": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Serial"
                    }
                },
                "error": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "types.Serial": {
            "type": "object",
            "properties": {
                "baud_rate": {
                    "description": "9600 115200",
                    "type": "integer"
                },
                "created": {
                    "description": "创建时间",
                    "type": "string"
                },
                "data_bits": {
                    "description": "5 6 7 8",
                    "type": "integer"
                },
                "desc": {
                    "description": "说明",
                    "type": "string"
                },
                "disabled": {
                    "type": "boolean"
                },
                "heartbeat": {
                    "description": "心跳包",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "parity_mode": {
                    "description": "0 1 2 NONE ODD EVEN",
                    "type": "integer"
                },
                "poller_interval": {
                    "description": "采集间隔",
                    "type": "integer"
                },
                "poller_period": {
                    "description": "采集周期",
                    "type": "integer"
                },
                "port_name": {
                    "description": "port, e.g. COM1 \"/dev/ttySerial1\".",
                    "type": "string"
                },
                "protocol_name": {
                    "description": "协议 rtu tcp parallel-tcp",
                    "type": "string"
                },
                "protocol_options": {
                    "description": "协议参数",
                    "type": "string"
                },
                "retry_maximum": {
                    "description": "最大次数",
                    "type": "integer"
                },
                "retry_timeout": {
                    "description": "重试时间",
                    "type": "integer"
                },
                "running": {
                    "type": "boolean"
                },
                "stop_bits": {
                    "description": "1 2",
                    "type": "integer"
                }
            }
        },
...
    }
}